+++ VSPController/VSPControllerBridge.h
@@ -35,7 +35,7 @@ bool IsDriverConnected(void);
 /**
  *
  */
-bool CreatePort(uint8_t baudRate, uint8_t dataBits, uint8_t stopBits, uint8_t parity, uint8_t flowCtrl);
+bool CreatePort(uint32_t baudRate, uint8_t dataBits, uint8_t stopBits, uint8_t parity, uint8_t flowCtrl);
 
 /**
  *
